// src/app/models/dto/user-details.dto.ts
import { Certification } from "../util/certification";
import { Education } from "../util/education";
import { Experience } from "../util/experience";
import { Skills } from "../util/skills";

export interface UserDetailsDto {
    first_name: string;
    last_name?: string;
    email: string;
    photo_url?: string;
    education: Education[];
    experience: Experience[];
    skills?: Skills;
    certifications: Certification[];
    achievements: string[];
    leetcode_url?: string;
    linkedin_url?: string;
    portfolio_url?: string;
    created_at: number;
    modified_at: number;
}